Improved AttemptToConnect with autoconnect and retries
The client will now attempt to autoconnect to a slave if IPAddress, TcpPort and SlaveId are found in appsettings.json.
If not, a new command. 'CONNECT' has been added to allow manual connection to a slave.

AttemptToConnect has now also got a retrycount to prevent it from being stuck in connecting state if unable to connect to a slave